requirements.txt
django==1.9
djangorestframework==3.3
django-filter==1.1.0